requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
coverage==6.2
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
mccabe==0.6.1
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
drf-yasg==1.20.0
